+++ scripts/tmfx.js
@@ -1,6 +1,6 @@
 export async function applyDDTint(placeable, color) {
   placeable = placeable.object ?? placeable;
-  color = PIXI.utils.string2hex(color);
+  color = _string2hex(color);
   if (isNaN(color)) {
     await TokenMagic.deleteFilters(placeable, 'DDTint');
   } else {
@@ -8,7 +8,7 @@ export async function applyDDTint(placeable, color) {
       {
         filterType: 'ddTint',
         filterId: 'DDTint',
-        tint: PIXI.utils.hex2rgb(color),
+        tint: _hex2string(color),
       },
     ]);
   }
@@ -21,7 +21,7 @@ export function getDDTint(placeable) {
     const filter = obj._TMFXgetSprite()?.filters?.find((f) => f.filterId === 'DDTint');
     if (filter) color = PIXI.utils.rgb2hex(filter.uniforms.tint);
   }
-  return PIXI.utils.hex2string(color);
+  return _hex2string(color);
 }
 
 export async function applyTMFXPreset(placeable, presetName, remove = false) {
@@ -34,5 +34,21 @@ export async function applyTMFXPreset(placeable, presetName, remove = false) {
   } else {
     const preset = TokenMagic.getPreset(presetName);
     if (preset) await TokenMagic.addUpdateFilters(placeable, deepClone(preset));
+  }
+}
+
+function _hex2string(color) {
+  if (PIXI.Color) {
+    return new PIXI.Color(color).toHex();
+  } else {
+    return PIXI.utils.hex2string(color);
+  }
+}
+
+function _string2hex(color) {
+  if (PIXI.Color) {
+    return new PIXI.Color(color).toNumber();
+  } else {
+    return PIXI.utils.string2hex(color);
   }
 }